+++ CHANGELOG.md
@@ -13,3 +13,6 @@
 ## 2026-02-10T02:06:51.179270+00:00
 - Workflows API endpoints: workflow-marketplace_patch_2026-02-10T02-06-42.441084+00-00.diff
 
+## 2026-02-10T03:22:27.183977+00:00
+- Workflows API endpoints: workflow-marketplace_patch_2026-02-10T03-22-18.260501+00-00.diff
+